local and Infura-connected tests split
