+++ witness/witness_gen_test.go
@@ -1547,103 +1547,6 @@ func TestDeleteToEmptyTrie(t *testing.T) {
 	GenerateProof("DeleteToEmptyTrie", trieModifications, statedb)
 }
 
-/*
-func TestFoo(t *testing.T) {
-	ks := [...]common.Hash{common.HexToHash("0x12"), common.HexToHash("0x21")}
-	var values []common.Hash
-	for i := 0; i < len(ks); i++ {
-		values = append(values, common.BigToHash(big.NewInt(int64(i + 1)))) // don't put 0 value because otherwise nothing will be set (if 0 is prev value), see state_object.go line 279
-	}
-	addr := common.HexToAddress("0x4E5B2e1dc63F6b91cb6Cd759936495434C7e972F")
-	var addresses []common.Address
-	for i := 0; i < len(ks); i++ {
-		addresses = append(addresses, addr)
-	}
-
-	nodeUrl := "https://mainnet.infura.io/v3/9aa3d95b3bc440fa88ea12eaa4456161"
-	blockNum := 14359865
-
-	GetProof(nodeUrl, blockNum, ks[:], values, addresses)
-}
-*/
-
-/*
-func TestFindAccount(t *testing.T) {
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-	
-	for i := 14000; i < 17000; i++ {
-		h := fmt.Sprintf("0x%d", i)
-		addr := common.HexToAddress(h)
-		// statedb.IntermediateRoot(false)
-		// statedb.CreateAccount(addr)
-
-		if statedb.GetCode(addr) == nil {
-			continue
-		}
-		accountProof, _, _, _, err := statedb.GetProof(addr)
-		fmt.Println(len(accountProof))
-		check(err)
-		fmt.Println(len(accountProof))
-		if len(accountProof) < 2 {
-			fmt.Println(len(accountProof))
-			fmt.Println("asdfsadf")
-		}
-
-		for i := 0; i < 1000; i++ {
-			h := fmt.Sprintf("0x2111d%d", i)
-			key2 := common.HexToHash(h)
-			val1 := common.BigToHash(big.NewInt(int64(1)))
-			statedb.SetState(addr, key2, val1)
-			statedb.IntermediateRoot(false)
-
-			storageProof, _, _, _, err := statedb.GetStorageProof(addr, key2)
-			check(err)
-			fmt.Println(len(storageProof))
-
-			v := common.Hash{} // empty value deletes the key
-			statedb.SetState(addr, key2, v)
-			statedb.IntermediateRoot(false)
-		}
-	}
-}
-*/
-
-/*
-func TestExtensionThreeBytesSel2(t *testing.T) {
-	// still searching for the right values
-	a := 0
-	h := fmt.Sprintf("0xf8a%d", a)
-	ks := []common.Hash{common.HexToHash(h)}
-	for i := 0; i < 1000; i++ {
-		a += 1
-		h := fmt.Sprintf("0xf8a%d", a)
-		ks = append(ks, common.HexToHash(h))
-	}
-	
-	var values []common.Hash
-	for i := 0; i < len(ks); i++ {
-		values = append(values, common.BigToHash(big.NewInt(int64(i + 1)))) // don't put 0 value because otherwise nothing will be set (if 0 is prev value), see state_object.go line 279
-	}
-
-	toBeModified := common.HexToHash("0xfa935")
-	addr := common.HexToAddress("0x75fbef2150818c32b36c57957226df4e24eb81c9")
-	val := common.BigToHash(big.NewInt(int64(17)))
-	trieMod := TrieModification{
-    	Type: StorageMod,
-		Key: toBeModified,
-		Value: val,
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	UpdateStateAndGenProof("ExtensionThreeBytesSel2", ks[:], values, addresses, trieModifications)
-}
-*/
-
 func TestUpdateTwoModifications(t *testing.T) {
 	ks := [...]common.Hash{common.HexToHash("0x12"), common.HexToHash("0x21")}
 	var values []common.Hash
@@ -1676,31 +1579,6 @@ func TestUpdateTwoModifications(t *testing.T) {
 
 	UpdateStateAndGenProof("UpdateTwoModifications", ks[:], values, addresses, trieModifications)
 }
-
-/*
-func TestFindAccountWithPlaceholderBranch(t *testing.T) {
-	blockNum := 13284469
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-	
-	for i := 0; i < 100; i++ {
-		h := fmt.Sprintf("0x%d", i)
-		addr := common.HexToAddress(h)
-		oracle.PrefetchAccount(statedb.Db.BlockNumber, addr, nil)
-
-		accountProof, _, _, _, _ := statedb.GetProof(addr)
-		statedb.CreateAccount(addr)
-		statedb.IntermediateRoot(false)
-		accountProof1, _, _, _, _ := statedb.GetProof(addr)
-
-		if len(accountProof1) == len(accountProof) + 1 {
-			fmt.Println("a;lskdfja;slkdfj")
-		}
-	}
-}
-*/
 
 func TestNonceModCShort(t *testing.T) {
 	blockNum := 14766377
@@ -1984,267 +1862,6 @@ func TestNonExistingAccount(t *testing.T) {
 	GenerateProof("NonExistingAccount", trieModifications, statedb)
 }
 
-func TestNonExistingAccountNilObjectInFirstLevel(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	i := 21
-	h := fmt.Sprintf("0x%d", i)
-	addr := common.HexToAddress(h)
-
-	trieMod := TrieModification{
-		Address: addr,
-    	Type: NonExistingAccount,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof("NonExistingAccountNilObjectInFirstLevel", trieModifications, statedb)
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
-func TestNonExistingAccountInFirstLevel(t *testing.T) {
-	// Only one element in the trie - the account with "wrong" address.
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	i := 10
-	h := fmt.Sprintf("0x%d", i)
-	addr := common.HexToAddress(h)
-
-	trieMod := TrieModification{
-    	Type: NonExistingAccount,
-		Balance: big.NewInt(23),
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProofSpecial("NonExistingAccountInFirstLevel", trieModifications, statedb, 4)
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
-func TestNonExistingAccountAfterFirstLevel(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	i := 22
-	h := fmt.Sprintf("0x%d", i)
-	addr := common.HexToAddress(h)
-
-	trieMod := TrieModification{
-		Address: addr,
-    	Type: NonExistingAccount,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof("NonExistingAccountAfterFirstLevel", trieModifications, statedb)
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
-// Account leaf after one branch. No storage proof.
-func TestAccountAfterFirstLevel(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	i := 21
-	h := fmt.Sprintf("0x%d", i)
-	addr := common.HexToAddress(h)
-
-	trieMod := TrieModification{
-    	Type: BalanceMod,
-		Balance: big.NewInt(23),
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof("AccountAfterFirstLevel", trieModifications, statedb)
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
-// Account leaf in first level. No storage proof.
-func TestAccountInFirstLevel(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	i := 21
-	h := fmt.Sprintf("0x%d", i)
-	addr := common.HexToAddress(h)
-
-	trieMod := TrieModification{
-    	Type: NonceMod,
-		Balance: big.NewInt(23),
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProofSpecial("AccountInFirstLevel", trieModifications, statedb, 1)
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
-func TestAccountExtensionInFirstLevel(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	h := fmt.Sprintf("0xa21%d", 0)
-	addr := common.HexToAddress(h)
-	found := false
-	for i := 0; i < 100000; i++ {
-		h := fmt.Sprintf("0xa21%d", i)
-		addr = common.HexToAddress(h)
-
-		statedb.CreateAccount(addr)
-		statedb.IntermediateRoot(false)
-
-		oracle.PrefetchAccount(statedb.Db.BlockNumber, addr, nil)
-		proof1, _, _, _, err := statedb.GetProof(addr)
-		check(err)
-
-		for j := 0; j < len(proof1) - 1; j++ {
-			if proof1[j][0] < 248 { // searching extension node
-				found = true
-			} 
-		}	
-
-		if found {
-			break
-		}
-	}
-
-	trieMod := TrieModification{
-    	Type: NonceMod,
-		Balance: big.NewInt(23),
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProofSpecial("AccountExtensionInFirstLevel", trieModifications, statedb, 5)
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
-func TestAccountBranchPlaceholder(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	h := fmt.Sprintf("0xab%d", 0)
-	addr := common.HexToAddress(h)
-	// Implicitly create account such that the account from the first level will be
-	// replaced by a branch.
-	trieMod := TrieModification{
-    	Type: NonceMod,
-		Balance: big.NewInt(23),
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProofSpecial("AccountBranchPlaceholder", trieModifications, statedb, 2) // don't use the same number as in the test below ("*PlaceholderInFirstLevel")
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
-func TestAccountBranchPlaceholderInFirstLevel(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	/*
-	for i := 0; i < 100000; i++ {
-		h := fmt.Sprintf("0xa21%d", i)
-		addr := common.HexToAddress(h)
-
-		oracle.PrefetchAccount(statedb.Db.BlockNumber, addr, nil)
-		proof1, _, _, _, err := statedb.GetProof(addr)
-		check(err)
-
-		statedb.CreateAccount(addr)
-		statedb.IntermediateRoot(false)
-
-		// addrHash1 := crypto.Keccak256Hash(addr.Bytes())
-		// addrHash1[31] = (addrHash1[31] + 1) % 255 // just some change
-
-		// oracle.PrefetchAccount(statedb.Db.BlockNumber, addr1, nil)
-		// proof11, _, _, err := statedb.GetProofByHash(common.BytesToHash(addrHash1.Bytes()))
-
-		statedb.CreateAccount(addr)
-		statedb.IntermediateRoot(false)
-
-		proof2, _, _, _, err := statedb.GetProof(addr)
-		check(err)
-		if len(proof1) + 1 == len(proof2) && len(proof1) == 1 {
-			elems, _, err := rlp.SplitList(proof1[len(proof1)-1])
-			if err != nil {
-				fmt.Println("decode error", err)
-			}
-			switch c, _ := rlp.CountValues(elems); c {
-			case 2:
-				fmt.Println("2")
-			case 17:
-			default:
-				fmt.Println("invalid number of list elements")
-			}
-		}
-	}
-	*/
-
-	h := fmt.Sprintf("0xab%d", 0)
-	addr := common.HexToAddress(h)
-	// Implicitly create account such that the account from the first level will be
-	// replaced by a branch.
-	trieMod := TrieModification{
-    	Type: BalanceMod,
-		Balance: big.NewInt(23),
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProofSpecial("AccountBranchPlaceholderInFirstLevel", trieModifications, statedb, 3) // don't use the same number as in the test above
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
 // Account proof after placeholder branch deeper in the trie (branch placeholder not in the
 // first or second level).
 func TestAccountBranchPlaceholderDeeper(t *testing.T) {
@@ -2266,173 +1883,6 @@ func TestAccountBranchPlaceholderDeeper(t *testing.T) {
 	trieModifications := []TrieModification{trieMod}
 
 	GenerateProofSpecial("AccountBranchPlaceholderDeeper", trieModifications, statedb, 0)
-}
-
-func TestStorageInFirstAccountInFirstLevel(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	i := 21
-	h := fmt.Sprintf("0x%d", i)
-	addr := common.HexToAddress(h)
-
-	trieMod := TrieModification{
-    	Type: StorageMod,
-		Key: common.HexToHash("0x12"),
-		Value: common.BigToHash(big.NewInt(int64(17))),
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProofSpecial("StorageInFirstAccountInFirstLevel", trieModifications, statedb, 1)
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
-func TestExtensionTwoNibblesInEvenLevel(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	h := fmt.Sprintf("0xa21%d", 0)
-	addr := common.HexToAddress(h)
-	found := false
-	for i := 0; i < 100000; i++ {
-		h := fmt.Sprintf("0xa21%d", i)
-		addr = common.HexToAddress(h)
-
-		statedb.CreateAccount(addr)
-		statedb.IntermediateRoot(false)
-
-		oracle.PrefetchAccount(statedb.Db.BlockNumber, addr, nil)
-		proof1, _, _, _, err := statedb.GetProof(addr)
-		check(err)
-
-		for j := 0; j < len(proof1) - 1; j++ {
-			if proof1[j][0] == 228 && proof1[j][1] == 130 && j % 2 == 0 {
-				fmt.Println(proof1[j])
-				found = true
-			} 
-		}	
-
-		if found {
-			break
-		}
-	}
-
-	trieMod := TrieModification{
-    	Type: NonceMod,
-		Nonce: 33,
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof("AccountExtensionTwoNibblesInEvenLevel", trieModifications, statedb)
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
-func TestExtensionThreeNibblesInEvenLevel(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	h := fmt.Sprintf("0xa21%d", 0)
-	addr := common.HexToAddress(h)
-	found := false
-	for i := 0; i < 100000; i++ {
-		h := fmt.Sprintf("0xa21%d", i)
-		addr = common.HexToAddress(h)
-
-		statedb.CreateAccount(addr)
-		statedb.IntermediateRoot(false)
-
-		oracle.PrefetchAccount(statedb.Db.BlockNumber, addr, nil)
-		proof1, _, _, _, err := statedb.GetProof(addr)
-		check(err)
-
-		for j := 0; j < len(proof1) - 1; j++ {
-			if proof1[j][0] == 228 && proof1[j][1] == 130 && j % 2 == 1 {
-				fmt.Println(proof1[j])
-				found = true
-			} 
-		}	
-
-		if found {
-			break
-		}
-	}
-
-	trieMod := TrieModification{
-    	Type: NonceMod,
-		Nonce: 33,
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof("AccountExtensionThreeNibblesInEvenLevel", trieModifications, statedb)
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
-func TestExtensionThreeNibblesInOddLevel(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	h := fmt.Sprintf("0xa21%d", 0)
-	addr := common.HexToAddress(h)
-	found := false
-	for i := 0; i < 100000; i++ {
-		h := fmt.Sprintf("0xa21%d", i)
-		addr = common.HexToAddress(h)
-
-		statedb.CreateAccount(addr)
-		statedb.IntermediateRoot(false)
-
-		oracle.PrefetchAccount(statedb.Db.BlockNumber, addr, nil)
-		proof1, _, _, _, err := statedb.GetProof(addr)
-		check(err)
-
-		for j := 0; j < len(proof1) - 1; j++ {
-			if proof1[j][0] == 228 && proof1[j][1] == 130 && proof1[j][2] != 0 && j % 2 == 0 {
-				fmt.Println(proof1[j])
-				found = true
-			} 
-		}	
-
-		if found {
-			break
-		}
-	}
-
-	trieMod := TrieModification{
-    	Type: NonceMod,
-		Nonce: 33,
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof("AccountExtensionThreeNibblesInOddLevel", trieModifications, statedb)
-
-	oracle.NodeUrl = oracle.RemoteUrl
 }
 
 func TestLeafInLastLevel(t *testing.T) {
@@ -2508,70 +1958,6 @@ func TestLeafInLastLevel(t *testing.T) {
 
 	oracle.PreventHashingInSecureTrie = false
 }
-
-/*
-	This makes the extension node longer than 55 bytes, so the first byte goes from 247 to 248.
-
-	storageProof[0]
-		[]uint8 len: 68, cap: 68, [248,66,160,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,160,172,62,32,109,202,28,158,228,207,196,4,119,200,12,209,177,23,17,228,25,192,126,235,26,91,197,92,194,...+4 more]
-	storageProof[1]
-		[]uint8 len: 83, cap: 83, [248,81,128,160,64,191,43,139,208,5,23,167,184,169,229,253,72,166,37,182,61,14,94,42,159,36,76,149,223,67,72,199,7,140,114,162,128,160,64,191,43,139,208,5,23,167,184,169,229,253,72,166,37,182,61,14,94,42,159,36,76,149,223,67,...+1...
-	storageProof[2]
-		[]uint8 len: 36, cap: 36, [227,32,161,160,187,239,170,18,88,1,56,188,38,60,149,117,120,38,223,78,36,235,129,201,170,170,170,170,170,170,170,170,170,170,170,170]
-
-	storageProof1[0]
-		[]uint8 len: 68, cap: 68, [248,66,160,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,160,104,147,154,48,225,9,124,91,65,60,63,202,4,255,242,29,212,241,132,56,213,231,51,133,235,124,255,172,...+4 more]
-	storageProof1[1]
-		[]uint8 len: 52, cap: 52, [243,128,194,32,17,128,160,64,191,43,139,208,5,23,167,184,169,229,253,72,166,37,182,61,14,94,42,159,36,76,149,223,67,72,199,7,140,114,162,128,128,128,128,128,128,128,128,128,128,128,128,128]
-	storageProof1[2]
-		[]uint8 len: 3, cap: 3, [194,32,17]
-*/
-/* TODO: extension nodes longer than 55 bytes yet to be implemented in the circuit (but the probability
-that they appear is very low).
-func TestLongExt(t *testing.T) {
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-	addr := common.HexToAddress("0x50efbf12580138bc623c95757286df4e24eb81c9")
-
-	statedb.DisableLoadingRemoteAccounts()
-	
-	statedb.CreateAccount(addr)
-
-	oracle.PreventHashingInSecureTrie = true
-
-	key1 := common.HexToHash("0x1")
-	
-	v1 := common.FromHex("0xbbefaa12580138bc263c95757826df4e24eb81c9aaaaaaaaaaaaaaaaaaaaaaaa")
-	val1 := common.BytesToHash(v1)
-
-	statedb.SetState(addr, key1, val1)
-
-	key2 := common.HexToHash("0x3")
-	statedb.SetState(addr, key2, val1)
-	statedb.IntermediateRoot(false)
-
-	storageProof, _, _, _, err := statedb.GetStorageProof(addr, key1)
-	check(err)
-
-	fmt.Println(storageProof[0])
-
-	val := common.BigToHash(big.NewInt(int64(17)))
-	trieMod := TrieModification{
-    	Type: StorageMod,
-		Key: key1,
-		Value: val,
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof("LongExt", trieModifications, statedb)
-
-	oracle.PreventHashingInSecureTrie = false
-}
-*/
 
 func TestLeafWithOneNibble(t *testing.T) {
 	blockNum := 0
@@ -2910,53 +2296,6 @@ func TestBranchAfterExtNode(t *testing.T) {
 	oracle.PreventHashingInSecureTrie = false
 }
 
-/*
-// Extension node going from [247,...] to [248,...].
-func TestExtNodeDiffLength(t *testing.T) {
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-	addr := common.HexToAddress("0x50efbf12580138bc623c95757286df4e24eb81c9")
-
-	statedb.DisableLoadingRemoteAccounts()
-	
-	statedb.CreateAccount(addr)
-
-	oracle.PreventHashingInSecureTrie = true // to store the unchanged key
-
-	key1 := common.HexToHash("0x10")
-	val1 := common.BigToHash(big.NewInt(int64(1)))
-
-	statedb.SetState(addr, key1, val1)
-
-	key2 := common.HexToHash("0x30")
-	statedb.SetState(addr, key2, val1)
-	statedb.IntermediateRoot(false)
-
-	storageProof, _, _, _, err := statedb.GetStorageProof(addr, key1)
-	check(err)
-
-	fmt.Println(storageProof[0])
-
-	// make the value long to have different kind of RLP in C proof
-	v1 := common.FromHex("0xbbefaa12580138bc263c95757826df4e24eb81c9aaaaaaaaaaaaaaaaaaaaaaaa")
-	val := common.BytesToHash(v1)
-	trieMod := TrieModification{
-    	Type: StorageMod,
-		Key: key1,
-		Value: val,
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof("ExtNodeDiffLength", trieModifications, statedb)
-
-	oracle.PreventHashingInSecureTrie = false
-}
-*/
-
 func TestNonExistingStorage(t *testing.T) {
 	ks := [...]common.Hash{common.HexToHash("0x11"), common.HexToHash("0x12")}
 	// hexed keys:
@@ -3012,65 +2351,6 @@ func TestNonExistingStorageLong(t *testing.T) {
 	UpdateStateAndGenProof("NonExistingStorageLong", ks[:], values, addresses, trieModifications)
 }
 
-func TestStorageInFirstLevelNonExisting(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	i := 21
-	h := fmt.Sprintf("0x%d", i)
-	addr := common.HexToAddress(h)
-
-	val1 := common.BigToHash(big.NewInt(int64(1)))
-	statedb.SetState(addr, common.HexToHash("0x11"), val1)
-	statedb.IntermediateRoot(false)
-
-	trieMod := TrieModification{
-    	Type: NonExistingStorage,
-		Key: common.HexToHash("0x12"),
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof("StorageInFirstLevelNonExisting", trieModifications, statedb)
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
-func TestStorageInFirstLevelNonExistingLong(t *testing.T) {
-	// geth --dev --http --ipcpath ~/Library/Ethereum/geth.ipc
-	oracle.NodeUrl = oracle.LocalUrl
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-
-	i := 21
-	h := fmt.Sprintf("0x%d", i)
-	addr := common.HexToAddress(h)
-
-	v1 := common.FromHex("0xbbefaa12580138bc263c95757826df4e24eb81c9aaaaaaaaaaaaaaaaaaaaaaaa")
-	val1 := common.BytesToHash(v1)
-	statedb.SetState(addr, common.HexToHash("0x11"), val1)
-	statedb.IntermediateRoot(false)
-
-	trieMod := TrieModification{
-    	Type: NonExistingStorage,
-		Key: common.HexToHash("0x12"),
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof("StorageInFirstLevelNonExistingLong", trieModifications, statedb)
-
-	oracle.NodeUrl = oracle.RemoteUrl
-}
-
 func TestNonExistingStorageNil(t *testing.T) {
 	// Nil in branch.
 	ks := [...]common.Hash{common.HexToHash("0x11"), common.HexToHash("0x12")}
@@ -3100,297 +2380,4 @@ func TestNonExistingStorageNil(t *testing.T) {
 	trieModifications := []TrieModification{trieMod}
 
 	UpdateStateAndGenProof("NonExistingStorageNil", ks[:], values, addresses, trieModifications)
-}
-
-func ExtNodeInserted(key1, key2, key3 common.Hash, testName string) {
-	// until infura is back up:
-	oracle.NodeUrl = oracle.LocalUrl
-
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-	addr := common.HexToAddress("0x50efbf12580138bc623c95757286df4e24eb81c9")
-
-	statedb.DisableLoadingRemoteAccounts()
-	statedb.CreateAccount(addr)
-	oracle.PreventHashingInSecureTrie = true // to store the unchanged key
-
-	// make the value long to have a hashed branch
-	v1 := common.FromHex("0xbbefaa12580138bc263c95757826df4e24eb81c9aaaaaaaaaaaaaaaaaaaaaaaa")
-	val1 := common.BytesToHash(v1)
-	statedb.SetState(addr, key1, val1)
-
-	statedb.SetState(addr, key2, val1)
-	statedb.IntermediateRoot(false)
-
-	v1 = common.FromHex("0xbb")
-	val := common.BytesToHash(v1)
-	trieMod := TrieModification{
-    	Type: StorageMod,
-		Key: key3,
-		Value: val,
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof(testName, trieModifications, statedb)
-
-	oracle.PreventHashingInSecureTrie = false
-}
-
-func ExtNodeDeleted(key1, key2, key3 common.Hash, testName string) {
-	// until infura is back up:
-	oracle.NodeUrl = oracle.LocalUrl
-
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-	addr := common.HexToAddress("0x50efbf12580138bc623c95757286df4e24eb81c9")
-
-	statedb.DisableLoadingRemoteAccounts()
-	statedb.CreateAccount(addr)
-	oracle.PreventHashingInSecureTrie = true // to store the unchanged key
-
-	// make the value long to have a hashed branch
-	v1 := common.FromHex("0xbbefaa12580138bc263c95757826df4e24eb81c9aaaaaaaaaaaaaaaaaaaaaaaa")
-	val1 := common.BytesToHash(v1)
-	statedb.SetState(addr, key1, val1)
-	statedb.SetState(addr, key2, val1)
-	statedb.SetState(addr, key3, val1)
-
-	statedb.IntermediateRoot(false)
-
-	val := common.Hash{} // empty value deletes the key
-	trieMod := TrieModification{
-    	Type: StorageMod,
-		Key: key3,
-		Value: val,
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof(testName, trieModifications, statedb)
-
-	oracle.PreventHashingInSecureTrie = false
-}
-
-func TestExtNodeInsertedBefore6After1FirstLevel(t *testing.T) {
-	key1 := common.HexToHash("0x1234561000000000000000000000000000000000000000000000000000000000")
-	// key1 bytes: [1 * 16 + 2, 3 * 16 + 4, 5 * 16 + 6, 1 * 16, 0, ..., 0]
-
-	key2 := common.HexToHash("0x1234563000000000000000000000000000000000000000000000000000000000")
-	// key2 bytes: [1 * 16 + 2, 3 * 16 + 4, 5 * 16 + 6, 3 * 16, 0, ..., 0]
-	// We now have an extension node with nibbles: [1, 2, 3, 4, 5, 6].
-
-	key3 := common.HexToHash("0x1234400000000000000000000000000000000000000000000000000000000000")
-	
-	ExtNodeInserted(key1, key2, key3, "ExtNodeInsertedBefore6After1FirstLevel")
-}
-
-func TestExtNodeDeletedBefore6After1FirstLevel(t *testing.T) {
-	key1 := common.HexToHash("0x1234561000000000000000000000000000000000000000000000000000000000")
-	// key1 bytes: [1 * 16 + 2, 3 * 16 + 4, 5 * 16 + 6, 1 * 16, 0, ..., 0]
-
-	key2 := common.HexToHash("0x1234563000000000000000000000000000000000000000000000000000000000")
-	// key2 bytes: [1 * 16 + 2, 3 * 16 + 4, 5 * 16 + 6, 3 * 16, 0, ..., 0]
-	// We now have an extension node with nibbles: [1, 2, 3, 4, 5, 6].
-
-	key3 := common.HexToHash("0x1234400000000000000000000000000000000000000000000000000000000000")
-	
-	ExtNodeDeleted(key1, key2, key3, "ExtNodeDeletedBefore6After1FirstLevel")
-}
-
-func TestExtNodeInsertedBefore6After2FirstLevel(t *testing.T) {
-	key1 := common.HexToHash("0x1234561000000000000000000000000000000000000000000000000000000000")
-	// key1 bytes: [1 * 16 + 2, 3 * 16 + 4, 5 * 16 + 6, 1 * 16, 0, ..., 0]
-
-	key2 := common.HexToHash("0x1234563000000000000000000000000000000000000000000000000000000000")
-	// key2 bytes: [1 * 16 + 2, 3 * 16 + 4, 5 * 16 + 6, 3 * 16, 0, ..., 0]
-
-	key3 := common.HexToHash("0x1235400000000000000000000000000000000000000000000000000000000000")
-	// key3 bytes: [1 * 16 + 2, 3 * 16 + 5, 4 * 16 + 0, 0, ..., 0]
-
-	ExtNodeInserted(key1, key2, key3, "ExtNodeInsertedBefore6After2FirstLevel")	
-}
-
-func TestExtNodeInsertedBefore6After4FirstLevel(t *testing.T) {
-	key1 := common.HexToHash("0x1234561000000000000000000000000000000000000000000000000000000000")
-	// key1 bytes: [1 * 16 + 2, 3 * 16 + 4, 5 * 16 + 6, 1 * 16, 0, ..., 0]
-
-	key2 := common.HexToHash("0x1234563000000000000000000000000000000000000000000000000000000000")
-	// key2 bytes: [1 * 16 + 2, 3 * 16 + 4, 5 * 16 + 6, 3 * 16, 0, ..., 0]
-
-	key3 := common.HexToHash("0x1635400000000000000000000000000000000000000000000000000000000000")
-
-	ExtNodeInserted(key1, key2, key3, "ExtNodeInsertedBefore6After4FirstLevel")	
-}
-
-func TestExtNodeInsertedBefore5After3FirstLevel(t *testing.T) {
-	key1 := common.HexToHash("0x2345610000000000000000000000000000000000000000000000000000000000")
-	key2 := common.HexToHash("0x2345630000000000000000000000000000000000000000000000000000000000")
-	key3 := common.HexToHash("0x2635400000000000000000000000000000000000000000000000000000000000")
-	
-	ExtNodeInserted(key1, key2, key3, "ExtNodeInsertedBefore5After3FirstLevel")	
-}
-
-func TestExtNodeInsertedBefore5After2FirstLevel(t *testing.T) {
-	key1 := common.HexToHash("0x2345610000000000000000000000000000000000000000000000000000000000")
-	key2 := common.HexToHash("0x2345630000000000000000000000000000000000000000000000000000000000")
-	key3 := common.HexToHash("0x2335400000000000000000000000000000000000000000000000000000000000")
-
-	ExtNodeInserted(key1, key2, key3, "ExtNodeInsertedBefore5After2FirstLevel")	
-}
-
-func TestExtNodeInsertedBefore5After1FirstLevel(t *testing.T) {
-	key1 := common.HexToHash("0x2345610000000000000000000000000000000000000000000000000000000000")
-	key2 := common.HexToHash("0x2345630000000000000000000000000000000000000000000000000000000000")
-	key3 := common.HexToHash("0x2343540000000000000000000000000000000000000000000000000000000000")
-
-	ExtNodeInserted(key1, key2, key3, "ExtNodeInsertedBefore5After1FirstLevel")	
-}
-
-func TestExtNodeInsertedBefore4After1(t *testing.T) {
-	oracle.NodeUrl = oracle.LocalUrl
-
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-	addr := common.HexToAddress("0x50efbf12580138bc623c95757286df4e24eb81c9")
-
-	statedb.DisableLoadingRemoteAccounts()
-	
-	statedb.CreateAccount(addr)
-
-	oracle.PreventHashingInSecureTrie = true // to store the unchanged key
-
-	val0 := common.BigToHash(big.NewInt(int64(1)))
-	key0 := common.HexToHash("0x1000000000000000000000000000000000000000000000000000000000000000")
-	statedb.SetState(addr, key0, val0)
-
-	key00 := common.HexToHash("0x0000000000000000000000000000000000000000000000000000000000000000")
-	statedb.SetState(addr, key00, val0)
-
-	key1 := common.HexToHash("0x1234561000000000000000000000000000000000000000000000000000000000")
-	// key1 bytes: [1 * 16 + 2, 3 * 16 + 4, 5 * 16 + 6, 1 * 16, 0, ..., 0]
-
-	// make the value long to have a hashed branch
-	v1 := common.FromHex("0xbbefaa12580138bc263c95757826df4e24eb81c9aaaaaaaaaaaaaaaaaaaaaaaa")
-	val1 := common.BytesToHash(v1)
-	// val1 := common.BigToHash(big.NewInt(int64(1)))
-	statedb.SetState(addr, key1, val1)
-
-	key2 := common.HexToHash("0x1234563000000000000000000000000000000000000000000000000000000000")
-	// key2 bytes: [1 * 16 + 2, 3 * 16 + 4, 5 * 16 + 6, 3 * 16, 0, ..., 0]
-
-	// We now have an extension node with nibbles: [3, 4, 5, 6].
-
-	statedb.SetState(addr, key2, val1)
-	statedb.IntermediateRoot(false)
-
-	key3 := common.HexToHash("0x1234400000000000000000000000000000000000000000000000000000000000")
-	// After adding key3 we will have an extension node with nibbles [3, 4]
-	// and another one with nibbles [5, 6].
-	
-	v1 = common.FromHex("0xbb")
-	val := common.BytesToHash(v1)
-	trieMod := TrieModification{
-    	Type: StorageMod,
-		Key: key3,
-		Value: val,
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof("ExtNodeInsertedBefore4After1", trieModifications, statedb)
-
-	oracle.PreventHashingInSecureTrie = false
-}
-
-func TestExtNodeDeletedBefore4After1(t *testing.T) {
-	oracle.NodeUrl = oracle.LocalUrl
-
-	blockNum := 0
-	blockNumberParent := big.NewInt(int64(blockNum))
-	blockHeaderParent := oracle.PrefetchBlock(blockNumberParent, true, nil)
-	database := state.NewDatabase(blockHeaderParent)
-	statedb, _ := state.New(blockHeaderParent.Root, database, nil)
-	addr := common.HexToAddress("0x50efbf12580138bc623c95757286df4e24eb81c9")
-
-	statedb.DisableLoadingRemoteAccounts()
-	
-	statedb.CreateAccount(addr)
-
-	oracle.PreventHashingInSecureTrie = true // to store the unchanged key
-
-	val0 := common.BigToHash(big.NewInt(int64(1)))
-	key0 := common.HexToHash("0x1000000000000000000000000000000000000000000000000000000000000000")
-	statedb.SetState(addr, key0, val0)
-
-	key00 := common.HexToHash("0x0000000000000000000000000000000000000000000000000000000000000000")
-	statedb.SetState(addr, key00, val0)
-
-	key1 := common.HexToHash("0x1234561000000000000000000000000000000000000000000000000000000000")
-
-	// make the value long to have a hashed branch
-	v1 := common.FromHex("0xbbefaa12580138bc263c95757826df4e24eb81c9aaaaaaaaaaaaaaaaaaaaaaaa")
-	val1 := common.BytesToHash(v1)
-	statedb.SetState(addr, key1, val1)
-
-	key2 := common.HexToHash("0x1234563000000000000000000000000000000000000000000000000000000000")
-	statedb.SetState(addr, key2, val1)
-
-	key3 := common.HexToHash("0x1234400000000000000000000000000000000000000000000000000000000000")
-	statedb.SetState(addr, key3, val1)
-	statedb.IntermediateRoot(false)
-	
-	val := common.Hash{} // empty value deletes the key
-	trieMod := TrieModification{
-    	Type: StorageMod,
-		Key: key3,
-		Value: val,
-		Address: addr,
-	}
-	trieModifications := []TrieModification{trieMod}
-
-	GenerateProof("ExtNodeDeletedBefore4After1", trieModifications, statedb)
-
-	oracle.PreventHashingInSecureTrie = false
-}
-
-func TestExtNodeInNewBranchFirstLevel(t *testing.T) {
-	key1 := common.HexToHash("0x2345610000000000000000000000000000000000000000000000000000000000")
-	key2 := common.HexToHash("0x2345630000000000000000000000000000000000000000000000000000000000")
-	key3 := common.HexToHash("0x6354000000000000000000000000000000000000000000000000000000000000")
-
-	ExtNodeInserted(key1, key2, key3, "ExtNodeInsertedInNewBranchFirstLevel")	
-}
-
-func TestExtNodeDeletedBranchDeletedFirstLevel(t *testing.T) {
-	key1 := common.HexToHash("0x2345610000000000000000000000000000000000000000000000000000000000")
-	key2 := common.HexToHash("0x2345630000000000000000000000000000000000000000000000000000000000")
-	key3 := common.HexToHash("0x6354000000000000000000000000000000000000000000000000000000000000")
-	
-	ExtNodeDeleted(key1, key2, key3, "ExtNodeDeletedBranchDeletedFirstLevel")
-}
-
-func TestExtNodeInsertedExtShortIsBranchFirstLevel(t *testing.T) {
-	key1 := common.HexToHash("0x2345610000000000000000000000000000000000000000000000000000000000")
-	key2 := common.HexToHash("0x2345630000000000000000000000000000000000000000000000000000000000")
-	key3 := common.HexToHash("0x2345100000000000000000000000000000000000000000000000000000000000")
-
-	ExtNodeInserted(key1, key2, key3, "ExtNodeInsertedExtShortIsBranchFirstLevel")	
-}
-
-func TestExtNodeDeletedExtShortIsBranchFirstLevel(t *testing.T) {
-	key1 := common.HexToHash("0x2345610000000000000000000000000000000000000000000000000000000000")
-	key2 := common.HexToHash("0x2345630000000000000000000000000000000000000000000000000000000000")
-	key3 := common.HexToHash("0x2345100000000000000000000000000000000000000000000000000000000000")
-
-	ExtNodeDeleted(key1, key2, key3, "ExtNodeInsertedExtShortIsBranchFirstLevel")	
 }
